Reorganised the models and added additional formatting to the videos
