+++ resolvers/Videos/MovieResolver.js
@@ -1,5 +1,5 @@
 const axios = require("axios");
-const { filter } = require("lodash");
+const { filter, forEach } = require("lodash");
 
 const { generateVideoEndpoint } = require("../../utils/generateEndpoints");
 
@@ -12,10 +12,12 @@ const MovieVideoResolver = async (parent, args, context, info) => {
     const { results } = data;
 
     // Filter to find the Trailers which are from YouTube
-    const YoutubeVideos = filter(
-      results,
-      video => video.type === "Trailer" && video.site === "YouTube"
-    );
+    const YoutubeVideos = filter(results, (video) => video.site === "YouTube");
+
+    // Add the url for each video
+    forEach(YoutubeVideos, (video) => {
+      video.url = `https://www.youtube.com/watch?v=${video.key}`;
+    });
 
     return YoutubeVideos;
   } catch (err) {
